[SurfaceFlinger] Filter out wide color modes for built-in display when necessary.
OEMs can choose whether to advertize the built-in display is wide color capable
regardless of whether hardware composer returns wide color modes or not. This
patch makes sure those this configuration is taken into account when we return
the color modes of the built-in display. Otherwise those devices will fail the
Display CTS tests.

BUG: 133010445
Test: Boot. Verified with Pixel 3.
Change-Id: Id70c39a4d274400f2c66b0a04728524051cb5c6a