Refine onPointerDownOutsideFocus
The original logic would send onPointerDownOutsideFocus to WindowManager
if a non-focused window is getting touched. And then, WindowManager will
move the display to the top. However, when per-display focus is enabled,
touching on a non-top focused window won't make the display to be moved
to the top. This makes the window cannot receive display-unspecified key
events, and also break some CTS.

This CL would send onPointerDownOutsideFocus if a window is getting
touched, as long as it is not the focused window on the focused display.

Fix: 134157810
Fix: 133323637
Test: atest ActivityLifecycleTopResumedStateTests WindowFocusTests
Change-Id: Ifa42ba766f58ec77e99f67080fd541284bb23c82