libtimeinstate: support cpufreq fast switching
In order to support kernels that use fast frequency switching, the
time_in_state BPF program needs to know which CPUs are members of
which cpufreq policies. Add logic to startTrackingUidCpuFreqTimes() to
write this information into a BPF map in order to make it available to
the BPF program.

Test: libtimeinstate_test passes
Change-Id: I47b38457766d21c2aa0879f156fc90757c0db705
Signed-off-by: Connor O'Brien <connoro@google.com>
Merged-In: I47b38457766d21c2aa0879f156fc90757c0db705