GpuStats: move the stats send at Activity launch off UI thread
Prevously when App Activity launches, we send GpuStats on the main UI thread.
The async binder to send the stats only takes 30us. However, to get the
gpuservice, the sync binder takes around 1.5ms because of the massive binders
on CPU at Activity launch. Thus we just move the stats sending to a separate
thread off the main UI thread. Now it's just 50us cost on the main UI thread.

Bug: 136228448
Test: systrace and check
Change-Id: Iff47dbc6a26a9d62ad33e92f93c079ac74fb3ea2