Use std::set instead of SortedVector
We only care whether a certain input device is present in a list or not.
For that, we can just use an std::set and not have to use custom
SortedVector data structure. This makes it easier for someone already
familiar with c++ but not necessarily Android to follow the code. It
also ensures that the data structure is properly maintained going
forward.

Bug: 137212522
Test: atest libinput_tests inputflinger_tests
Change-Id: I7808b41524ae0b87dbf1aaf7a49afa64b14b3eb9