installd: Account for empty  obb contents on migration
We need /data/media/obb on devices with old kernels. Since that folder
is present, we should check if there are contents there before migrating
data.

This also fixes the check for if the kernel supports the unshared_obb
option. If it does, it only does the migration if there is data to
migrate.

Test: Boot with GSI over P, boot normal Q, with and without data at
      /data/media/obb. With data, it should migrate. Without data it
      shoud skip the migration. On GSI on P it should skip the
      migration.
Bug: 136199978
Change-Id: I41b0ec7a8a6b57b2ea6a530630d56c00f758009a