[RenderEngine] Allow rounded corners shader to use lower precision.
Big corner radii would cause problems at lower precision because
computing a vector norm would overflow. But since highp tanks
performance, instead scale down and back up by a power of 2 during the
normalization operation, so that a higher precision is not required.

Bug: 137191934
Test: eyeballing systrace when opening Photos, GPU seems to take shorter
time and no jank is observed
Test: Stepping through frames from a Pixel 3 XL screenrecording shows
that the beginning of the app opening animation is rounded instead of
square

Change-Id: Ie7a39bd0f992f7f42414ef49de9873775cc34409