vkjson: correctly handle std::numeric_limits<float>::infinity()
Java JSON library can't handle infinity value. So if vkjson reports infinity
limit, we need to manually clamp the value within the Java JSON value range.
Up-casting float infinity will be equal to double infinity, and we need to make
sure the text representation is not going out of range between the cppjson to
Java json converison, so we have to clamp to a safe min/max range of double.

Bug: 134453786
Test: adb shell cmd gpu vkjson
Change-Id: I498fbcdb76ec55d6443ca202af191d6fb12fec12