SurfaceFlinger: get rid of LayerStats
LayerStats is not in use, so let's just remove it. If later SurfaceTracing has
been moved back to main thread, we could probably plumb HWC layer stats into
SurfaceTracing. For now, if we are interested in any statistical HWC stats, we
could add that into TimeStats so that we get the stats from the field as well.

Test: build, flash and boot
Change-Id: Iff418f60bc22d8867294f919699300cf25eee258