Do not report latency for injected events
When reporting the latency of touch events, ensure that injected events
are excluded. These events can have arbitrary timestamps, and would not
result in meaningful data. Also do not report events for statistics if
inputfilter is enabled.

Move the statistics reporting from InputTransport to InputDispatcher.
This ensures that there's only 1 instance of the mStatistics object.
This also provides easy access to the inputfilterenabled state.

Bug: 13894199
Test: Change the reporting period to 0 (to report every event immediately)
Inject events in various ways and ensure they don't go to statsd
$ m statsd_testdrive && ./out/host/linux-x86/bin/statsd_testdrive 34
$ adb shell input tap 100 100
$ adb shell monkey 1000
Next, relaunch the statsd_testdrive script and touch the screen
Observe that events are reported.

Change-Id: Ief8040599a347e084e75584ed3164c60a6dbc4ad